[WIP] Handling DTMF tones in a steaming phone call
Discussed here:
https://discord.com/channels/1079125925163180093/1105907413992882238/1106005912184967248

The goal is to do an outbound call where the bot will need to navigate a phone tree by listening to options and pressing keys on the dial pad.

I found wav files for DTMF here, which are free for all use including commercial:
https://freesound.org/people/ialexs/packs/10184/

I added a playground file to test this: play_dtmf.py.  This works for me.

I then added some logic into streaming_conversation.py to detect the intent to press a button from messages with a given format: `*presses 0*`.  If we detect that format, we play the sound.

When I tested this with a Twilio call, it didn't seem to work -- at least I couldn't hear the sound being played.

Also, I'm not sure this setup is very clean:
- We should give people a way to enable/disable this DTMF capability.  I'm not sure what's the best setup/config for this.
- We should give people a way to specify the format used to detect DTMF intent from the freeform GPT response.  Or maybe instead of using freeform responses, the agent should be responsible for providing entirely different _types_ of messages for DTMF?
- I decided to still proceed with text-to-speech, even after the DTMF tone is played.  Why?  Because then it still shows up in the transcript, which is convenient.  But it's still a little odd to hear the robot say "presses 2".