Add optional `digits` param to OutboundCall
When initiating an OutboundCall, this new parameter lets you specify digits (DTMF tones) to play at the start of the call, before the websocket is established and the vocode bot gets going.

The format is described here:
https://www.twilio.com/docs/voice/api/call-resource#create-a-call-resource

"A string of keys to dial after connecting to the number, maximum of 32 digits. Valid digits in the string include: any digit (0-9), '#', '*' and 'w', to insert a half second pause. For example, if you connected to a company phone number and wanted to pause for one second, and then dial extension 1234 followed by the pound key, the value of this parameter would be ww1234#. Remember to URL-encode this string, since the '#' character has special meaning in a URL. If both SendDigits and MachineDetection parameters are provided, then MachineDetection will be ignored."

This provides a rudimentary way to let the bot walk a phone tree before it proceeds to interact with a human on the other side.  On the plus side, it's simple and fast.  On other hand, it's very rigid.  It requires you to know, in advance, what the menu options are and what's the right sequence of keys to press.  It doesn't automatically adapt if the target phone line changes its IVR.

We're still discussing how to enable that "flexible" phone-tree bot (https://discord.com/channels/1079125925163180093/1110400160950927370/1110400163392016474), but that effort is facing the challenge of how to play a DTMF tone into the `<Stream>`, or how to `<Play>` a tone without interrupting the stream.  Until we crack that nut, at least we'll have this basic way to play DTMF tones at the start of the call.