[WIP] Play DTMF tones using Twilio `<Play>`
We have this ongoing effort to make the vocode bot capable of "walking phone trees" -- to listen to menu options, decide what to press, and then "press a button" (play DTMF tone).

Discussions here:
https://discord.com/channels/1079125925163180093/1110400160950927370/1110400163392016474

Our first attempt was to have wav files for each of those tones and to play the wav through the stream, as we might with typing-noise.wav.  Unfortunately, this doesn't work.  Twilio has confirmed that they don't support playing DTMF tones through the stream.

So, the new approach is to instead rely on Twilio's built-in `<Play>` verb.  You can pass it the `digits` to play.  This PR attempts to add that capability.

For the moment this approach _also_ doesn't seem to work.  When I update the call with `<Play>`, it immediately terminates the websocket with a "stop" message and ends the call.  As a workaround, I attempted to do a single update that includes both `<Play>` and `<Connect>` (to re-connect the stream) as suggested here:
https://stackoverflow.com/questions/73129566/sending-dtmf-through-a-stream/76325543#76325543
But that approach doesn't work either.  The call still terminates and there's no request to re-connect.

I'm working with Twilio support to diagnose this issue.  For now, I'm posting this WIP PR so other folks can look at my approach and comment / revise / test.