[BREAKING] Change the type of TranscriptCompleteEvent.transcript v3 (#139)
* [BREAKING] Change the type of TranscriptCompleteEvent.transcript

This is an alternative approach to the following PR:
https://github.com/vocodedev/vocode-python/pull/114

In this take, we use the type `Transcript` for the field `TranscriptCompleteEvent.transcript`.  We avoid circular dependency issues by moving the relevant event definitions from `events.py` into `transcript.py`.

This could be seen as beneficial, in that the transcript and related events are all contained in a single module.  It could also be seen as strange because we no longer have all events defined in a single module, and because the EventType and corresponding <Name>Event are defined in different places.

* move streaming/utils/transcript.py to streaming/models/transcript.py

As suggested by @ajar98