Remove CallsRouter.calls + save in-progress Transcript into config
The problem with the previous approach (saving the Call object into a hash in the CallsRouter) is that it only works with a one-server deployment.  If you have 2+ servers, then there's no guarantee that the second connection will go to the same server.  If the request goes to a different server, then that server won't have the Call object in the hash.

So, with this commit I'm going back to the setup where we create a brand new Call object each time we get a new websocket connection.  But, when we create that object from the call_config, we will pick up the in-progress Transcript if there is one.