Add LoggerAdapter earlier so it applies for all log messages
Previously, the unwrapped version of the logger was passed through to the transcriber/agent/synthesizer, so not all log messages included the conversation_id.  Now I'm initializing the conversation_id earlier, and using it to create the wrapped logger earlier, so now the conversation_id is used in all log messages.