Detect goodbye from the bot
Vocode has a flag to `end_conversation_on_goodbye` which uses GoodbyeModel to detect a "goodbye" intent in a message, and if so then to conclude the conversation.

However, it looks like this logic is only applied to human messages, not to bot messages.  This is a problem in my use case - I'm using vocode for outbound calls, which occasionally hit IVR systems that never say "goodbye" and never terminate, which means the call never terminates, and the poor vocode bot is stuck :-)

This PR applies the goodbye-detection logic to bot messages.

In practice this seems to work, though I do observe one side effect: I end up getting two TRANSCRIPT_COMPLETE events.

Worth thinking about:
1. How this would look in an action agent, where `end_call` is simply one of the actions the bot can take.
2. Besides ending a call via "goodbye" detection, there might be other rules for when to end a call.  For example, you might want to set a 1-hour limit on the call, and if vocode detects that it's been more than an hour it automatically ends the call (can be done without LLM).

Regardless, this PR gives people _some_ way to have the bot end the call.